pyproject.toml
[tool.poetry]
name = "deepfakecapstone"
version = "0.1.0"
description = "Capstone for Drake University's AI major"
authors = ["LaneAffield <laneaffield@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
randimage = "^1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
